# Cut a few seconds off of every test using Cockroach